Add HW key library
Add API to generate a hardware unique key (HUK)
for a device if one does not already exist.
This allows a device to have a secure way to
encrypt/decrypt data. This means only the device itself can access the data.
This is useful if a device wants to encrypt data at rest.

bug 21177